modularize tests a bit
