+++ test/test_running.py
@@ -1,21 +1,16 @@
 import json
-import os
 import random
 import re
 import shutil
 import string
 import subprocess
-import pathlib
 
 import requests
 import pytest
 
 #pylint:disable=redefined-outer-name,use-dict-literal,missing-timeout,unspecified-encoding,consider-using-with
 
-PROTO="http"
-HOST="localhost.pwn.college"
-CONTAINER_NAME = os.environ.get("CONTAINER_NAME", "dojo-test")
-TEST_DOJOS_LOCATION = pathlib.Path(__file__).parent / "dojos"
+from utils import CONTAINER_NAME, PROTO, HOST, login
 
 def dojo_run(*args, **kwargs):
     kwargs.update(stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
@@ -50,28 +45,6 @@ def get_challenge_id(session, dojo, module, challenge):
     return challenge_id
 
 
-def parse_csrf_token(text):
-    match = re.search("'csrfNonce': \"(\\w+)\"", text)
-    assert match, "Failed to find CSRF token"
-    return match.group(1)
-
-
-def login(name, password, *, success=True, register=False, email=None):
-    session = requests.Session()
-    endpoint = "login" if not register else "register"
-    nonce = parse_csrf_token(session.get(f"{PROTO}://{HOST}/{endpoint}").text)
-    data = dict(name=name, password=password, nonce=nonce)
-    if register:
-        data["email"] = email or f"{name}@example.com"
-    response = session.post(f"{PROTO}://{HOST}/{endpoint}", data=data, allow_redirects=False)
-    if not success:
-        assert response.status_code == 200, f"Expected {endpoint} failure (status code 200), but got {response.status_code}"
-        return session
-    assert response.status_code == 302, f"Expected {endpoint} success (status code 302), but got {response.status_code}"
-    session.headers["CSRF-Token"] = parse_csrf_token(session.get(f"{PROTO}://{HOST}/").text)
-    return session
-
-
 def start_challenge(dojo, module, challenge, practice=False, *, session):
     start_challenge_json = dict(dojo=dojo, module=module, challenge=challenge, practice=practice)
     response = session.post(f"{PROTO}://{HOST}/pwncollege_api/v1/docker", json=start_challenge_json)
@@ -88,33 +61,6 @@ def db_sql_one(sql):
 def get_user_id(user_name):
     return int(db_sql_one(f"SELECT id FROM users WHERE name = '{user_name}'"))
 
-@pytest.fixture(scope="module")
-def admin_session():
-    session = login("admin", "admin")
-    yield session
-
-
-@pytest.fixture
-def random_user():
-    random_id = "".join(random.choices(string.ascii_lowercase, k=16))
-    session = login(random_id, random_id, register=True)
-    yield random_id, session
-
-
-@pytest.fixture(scope="module")
-def completionist_user():
-    random_id = "".join(random.choices(string.ascii_lowercase, k=16))
-    session = login(random_id, random_id, register=True)
-    yield random_id, session
-
-
-@pytest.fixture(scope="module")
-def guest_dojo_admin():
-    random_id = "".join(random.choices(string.ascii_lowercase, k=16))
-    session = login(random_id, random_id, register=True)
-    yield random_id, session
-
-
 @pytest.mark.parametrize("endpoint", ["/", "/dojos", "/login", "/register"])
 def test_unauthenticated_return_200(endpoint):
     response = requests.get(f"{PROTO}://{HOST}{endpoint}")
@@ -130,25 +76,6 @@ def test_register():
     random_id = "".join(random.choices(string.ascii_lowercase, k=16))
     login(random_id, random_id, register=True)
 
-def make_dojo_official(dojo_rid, admin_session):
-    response = admin_session.post(f"{PROTO}://{HOST}/pwncollege_api/v1/dojo/{dojo_rid}/promote-dojo", json={})
-    assert response.status_code == 200, f"Expected status code 200, but got {response.status_code} - {response.json()}"
-
-def create_dojo(repository, *, session):
-    test_public_key = f"public/{repository}"
-    test_private_key = f"private/{repository}"
-    create_dojo_json = dict(repository=repository, public_key=test_public_key, private_key=test_private_key)
-    response = session.post(f"{PROTO}://{HOST}/pwncollege_api/v1/dojo/create", json=create_dojo_json)
-    assert response.status_code == 200, f"Expected status code 200, but got {response.status_code} - {response.json()}"
-    dojo_reference_id = response.json()["dojo"]
-    return dojo_reference_id
-
-def create_dojo_yml(spec, *, session):
-    response = session.post(f"{PROTO}://{HOST}/pwncollege_api/v1/dojo/create-spec", json={"spec": spec})
-    assert response.status_code == 200, f"Expected status code 200, but got {response.status_code} - {response.json()}"
-    dojo_reference_id = response.json()["dojo"]
-    return dojo_reference_id
-
 def start_and_solve(user_name, session, dojo, module, challenge):
     start_challenge(dojo, module, challenge, session=session)
     challenge_id = get_challenge_id(session, dojo, module, challenge)
@@ -159,34 +86,6 @@ def start_and_solve(user_name, session, dojo, module, challenge):
     )
     assert response.status_code == 200, f"Expected status code 200, but got {response.status_code}"
     assert response.json()["success"], "Expected to successfully submit flag"
-
-
-@pytest.fixture(scope="module")
-def example_dojo(admin_session):
-    rid = create_dojo("pwncollege/example-dojo", session=admin_session)
-    make_dojo_official(rid, admin_session)
-    return rid
-
-@pytest.fixture(scope="module")
-def belt_dojos(admin_session):
-    belt_dojo_rids = {
-        color: create_dojo_yml(
-            open(TEST_DOJOS_LOCATION / f"fake_{color}.yml").read(), session=admin_session
-        ) for color in [ "orange", "yellow", "green", "blue" ]
-    }
-    for rid in belt_dojo_rids.values():
-        make_dojo_official(rid, admin_session)
-    return belt_dojo_rids
-
-@pytest.fixture(scope="module")
-def example_import_dojo(admin_session):
-    rid = create_dojo("pwncollege/example-import-dojo", session=admin_session)
-    make_dojo_official(rid, admin_session)
-    return rid
-
-@pytest.fixture(scope="module")
-def simple_award_dojo(admin_session):
-    return create_dojo_yml(open(TEST_DOJOS_LOCATION / "simple_award_dojo.yml").read(), session=admin_session)
 
 
 @pytest.mark.dependency()
